fix: Indexable enum not being iterable error and move default properties of layouter out of class
This fix changes the index function to be a member function and not a static method which ensures we only pass in enum members. This also removes the need for the is_parsable function, so it is removed along with its tests. To keep the app clean, this also moves the default values out of the layouter class and thus the enums can be directly retrieved from the state with default fallback.